+++ apps/platform/api/src/database/schema.ts
@@ -66,7 +66,7 @@ export type PlatformAppsTable = {
   id: Generated<number>;
   label: string;
   module_key: string;
-  stack: "platform" | "billing" | "accounts" | "devkit" | "mail" | "platform-task-manager" | "blog";
+  stack: "platform" | "billing" | "accounts" | "devkit" | "mail" | "platform-task-manager" | "blog" | "auditor";
   updated_at: TimestampColumn;
   uuid: string;
 };