+++ apps/platform/api/src/modules/tenant/tenant.types.ts
@@ -9,7 +9,7 @@ export type Tenant = {
   dbType: string;
   dbUser: string;
   enabledModuleKeys: string[];
-  defaultLandingApp: "application" | "billing" | "accounts" | "devkit" | "mail" | "task-manager" | "blog";
+  defaultLandingApp: "application" | "billing" | "accounts" | "devkit" | "mail" | "task-manager" | "blog" | "auditor";
   id: number;
   mobile: string | null;
   payloadSettings: Record<string, unknown>;